Move shell integration addon to platform
